% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_worked.R
\name{add_minutes_worked}
\alias{add_minutes_worked}
\title{Add minutes worked between clocking in and out.}
\usage{
add_minutes_worked(.data)
}
\arguments{
\item{.data}{Clocking data from `.org` file.}
}
\value{
Add a column of minutes worked.
}
\description{
Add minutes worked between clocking in and out.
}
