% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_org_clock}
\alias{import_org_clock}
\title{Import clocking lines from .org files.}
\usage{
import_org_clock(file)
}
\arguments{
\item{file}{org file.}
}
\value{
Imported org file
}
\description{
Import clocking lines from .org files.
}
\examples{

import_org_clock("tests/testthat/clocking.org")

}
