% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worktime-package.R
\docType{package}
\name{worktime-package}
\alias{worktime}
\alias{worktime-package}
\title{worktime: Import and Visualize Hours Worked on Tasks}
\description{
Imports files that track minute-by-minute clocking of work time 
    (currently only org-mode files) and provides some plot functions to visualize
    various aspects of working or time tracked tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwjohnst86/worktime}
  \item Report bugs at \url{https://github.com/lwjohnst86/worktime/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com}

}
\keyword{internal}
